requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
alpaca-py>=0.21.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
